Mon, Jun 6, 2022, 7:42 PM +00:00

+++ readability/readability.c
@@ -80,7 +80,7 @@ int count_words(string text)
 }
 
 //counting sentences in text
-float count_sentences(string text)
+int count_sentences(string text)
 {
     int count_sentences = 0;
     for (int i = 0, n = strlen(text); i < n; i++)
